lending-origination/src/main/java/com/fusionx/lending/origination/resource/LeadInforBranchAssignUpdateResource.java
package com.fusionx.lending.origination.resource;

import com.fasterxml.jackson.databind.annotation.JsonSerialize;

import javax.validation.Valid;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Pattern;
import javax.validation.constraints.Size;
import java.util.List;

@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
public class LeadInforBranchAssignUpdateResource {

    @NotBlank(message = "{common.not-null}")
    @Pattern(regexp = "^$|[0-9]+", message = "{common-numeric.pattern}")
    private String branchId;

    @NotBlank(message = "{common.not-null}")
    @Size(max = 70, message = "{common-name.size}")
    private String branchName;

    @Valid
    private List<LeadListResource> leadIdList;


    public String getBranchName() {
        return branchName;
    }

    public void setBranchName(String branchName) {
        this.branchName = branchName;
    }

    public String getBranchId() {
        return branchId;
    }

    public void setBranchId(String branchId) {
        this.branchId = branchId;
    }

    public List<LeadListResource> getLeadIdList() {
        return leadIdList;
    }

    public void setLeadIdList(List<LeadListResource> leadIdList) {
        this.leadIdList = leadIdList;
    }
}
